experimento/cloud_function/requirements.txt
google-cloud-videointelligence
google-cloud-pubsub